requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
uvicorn[standard]
sqlalchemy>=2.0.0,<3.0.0
asyncpg>=0.29.0,<1.0.0
psycopg2-binary>=2.9.0,<3.0.0
alembic>=1.13.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=4.0.0,<5.0.0
email-validator>=2.0.0,<3.0.0
weasyprint>=60.0,<70.0
qrcode[pil]>=7.4.0,<8.0.0
Pillow>=10.0.0,<11.0.0
jinja2>=3.1.0,<4.0.0
